Cleanup styling code.
