+++ src/Avalonia.Styling/Styling/IStyleInstance.cs
@@ -14,6 +14,9 @@ namespace Avalonia.Styling
         /// </summary>
         IStyle Source { get; }
 
+        /// <summary>
+        /// Gets a value indicating whether this style is active.
+        /// </summary>
         bool IsActive { get; }
 
         /// <summary>
